requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.2
fonttools==4.54.1
frozendict==2.4.6
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
multitasking==0.0.11
numpy>=1.23.0,<2.0.0 
openai==0.28
packaging==24.2
pandas>=2.0.0,<2.1.0
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.4.0
seaborn==0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.40.0
tavily-python==0.5.0
tensorflow==2.18.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
webencodings==0.5.1
yfinance==0.2.44